Fix TrbSegment overflow bug

+++ src/terark/db/trbdb/trb_db_index.cpp
@@ -283,21 +283,21 @@ class TrbWritableIndexTemplate : public TrbWritableIndex
 
         fstring key(size_type i) const
         {
-            byte const *pos = data.at<data_object>(index[i].offset << index_shift).data;
+            byte const *pos = data.at<data_object>(size_t(index[i].offset) << index_shift).data;
             uint32_t len;
             FAST_READ_VAR_UINT32(pos, len);
             return fstring(pos, len);
         }
         byte const *key_ptr(size_type i) const
         {
-            byte const *pos = data.at<data_object>(index[i].offset << index_shift).data;
+            byte const *pos = data.at<data_object>(size_t(index[i].offset) << index_shift).data;
             uint32_t len;
             FAST_READ_VAR_UINT32(pos, len);
             return pos;
         }
         size_type key_len(size_type i) const
         {
-            byte const *pos = data.at<data_object>(index[i].offset << index_shift).data;
+            byte const *pos = data.at<data_object>(size_t(index[i].offset) << index_shift).data;
             uint32_t len;
             FAST_READ_VAR_UINT32(pos, len);
             return len;
@@ -361,7 +361,7 @@ class TrbWritableIndexTemplate : public TrbWritableIndex
             size_type dst_len = pool_type::align_to(d.size() + len_len);
             size_type pos = data.alloc(dst_len);
             assert(pos % 4 == 0);
-            index[i].offset = pos >> index_shift;
+            index[i].offset = uint32_t(pos >> index_shift);
             byte *dst_ptr = data.at<data_object>(pos).data;
             std::memcpy(dst_ptr, len_data, len_len);
             std::memcpy(dst_ptr + len_len, d.data(), d.size());
@@ -430,7 +430,7 @@ class TrbWritableIndexTemplate : public TrbWritableIndex
             size_type dst_len = pool_type::align_to(d.size() + len_len);
             size_type pos = data.alloc(dst_len);
             assert(pos % 4 == 0);
-            index[i].offset = pos >> index_shift;
+            index[i].offset = uint32_t(pos >> index_shift);
             byte *dst_ptr = data.at<data_object>(pos).data;
             std::memcpy(dst_ptr, len_data, len_len);
             std::memcpy(dst_ptr + len_len, d.data(), d.size());
@@ -453,7 +453,7 @@ class TrbWritableIndexTemplate : public TrbWritableIndex
                    )
                )
             {
-                data.sfree(index[i].offset << index_shift, dst_len);
+                data.sfree(size_t(index[i].offset) << index_shift, dst_len);
                 index[i].offset = index[c].offset;
             }
             total += d.size();
@@ -491,7 +491,7 @@ class TrbWritableIndexTemplate : public TrbWritableIndex
                     }
                 }
             }
-            byte const *ptr = data.at<data_object>(index[i].offset << index_shift).data, *end_ptr;
+            byte const *ptr = data.at<data_object>(size_t(index[i].offset) << index_shift).data, *end_ptr;
             size_type len = load_var_uint32(ptr, &end_ptr);
             if(true
                && (
@@ -506,7 +506,7 @@ class TrbWritableIndexTemplate : public TrbWritableIndex
                    )
                )
             {
-                data.sfree(index[i].offset << index_shift, pool_type::align_to(end_ptr - ptr + len));
+                data.sfree(size_t(index[i].offset) << index_shift, pool_type::align_to(end_ptr - ptr + len));
             }
             threaded_rbtree_remove(root,
                                    stack,
