+++ src/terark/db/trbdb/trb_db_store.cpp
@@ -206,7 +206,7 @@ fstring TrbWritableStore::readItem(size_type i) const
 {
     assert(i < m_index.size());
     byte const *ptr;
-    size_type len = load_var_uint32(m_data.at<data_object>(m_index[i] << index_shift).data, &ptr);
+    size_type len = load_var_uint32(m_data.at<data_object>(size_t(m_index[i]) << index_shift).data, &ptr);
     return fstring(ptr, len);
 }
 
@@ -226,7 +226,7 @@ void TrbWritableStore::storeItem(size_type i, fstring d)
     size_type dst_len = pool_type::align_to(d.size() + len_len);
     size_t pos = m_data.alloc(dst_len);
     assert(pos % 4 == 0);
-    m_index[i] = pos >> index_shift;
+    m_index[i] = uint32_t(pos >> index_shift);
     byte *dst_ptr = m_data.at<data_object>(pos).data;
     m_size += d.size();
     std::memcpy(dst_ptr, len_data, len_len);
@@ -236,7 +236,7 @@ void TrbWritableStore::storeItem(size_type i, fstring d)
 void TrbWritableStore::removeItem(size_type i)
 {
     assert(i < m_index.size() && m_index[i] != store_nil_index);
-    size_t pos = m_index[i] << index_shift;
+    size_t pos = size_t(m_index[i]) << index_shift;
     byte const *ptr = m_data.at<data_object>(pos).data, *end_ptr;
     size_type len = load_var_uint32(ptr, &end_ptr);
     m_data.sfree(pos, pool_type::align_to(end_ptr - ptr + len));
